refactor: move evaluation to own module
