+++ src/app/app.module.ts
@@ -3,17 +3,19 @@ import { BrowserModule } from '@angular/platform-browser';
 import { FormsModule } from '@angular/forms';
 import { AppComponent } from './app.component';
 import { SnackbarComponent } from './snackbar/snackbar.component';
-import { EvaluationComponent } from './evaluation/evaluation.component';
 import { StartListModule } from './start-list/start-list.module';
 import { TimekeepingModule } from './timekeeping/timekeeping.module';
+import { EvaluationModule } from './evaluation/evaluation.module';
 
 @NgModule({
-  declarations: [
-    AppComponent,
-    SnackbarComponent,
-    EvaluationComponent,
+  declarations: [AppComponent, SnackbarComponent],
+  imports: [
+    BrowserModule,
+    FormsModule,
+    StartListModule,
+    TimekeepingModule,
+    EvaluationModule,
   ],
-  imports: [BrowserModule, FormsModule, StartListModule, TimekeepingModule],
   providers: [],
   bootstrap: [AppComponent],
 })
